+++ packages/illustrations/src/__generated__/heroSquare/data/versionMap.ts
@@ -348,11 +348,12 @@ const versionMap: Record<HeroSquareName, number> = {
   sustainable: 0,
   gamer: 0,
   private: 0,
-  lightingnetworkTransfer: 0,
-  lightingnetworkInvoice: 0,
+  lightningNetworkTransfer: 0,
+  lightningNetworkInvoice: 0,
   realToUSDC: 0,
-  lightingnetwork: 0,
-  lightingnetworkSend: 0,
+  lightningNetwork: 0,
+  lightningNetworkSend: 0,
+  referralsGenericCoin: 0,
 };
 
 export default versionMap;