+++ packages/illustrations/src/__generated__/spotRectangle/data/svgJsMap.ts
@@ -413,9 +413,9 @@ const svgJsMap = {
     light: () => require('../svgJs/light/leverage-3').content,
     dark: () => require('../svgJs/dark/leverage-3').content,
   },
-  lightingnetworkSend: {
-    light: () => require('../svgJs/light/lightingnetworkSend-0').content,
-    dark: () => require('../svgJs/dark/lightingnetworkSend-0').content,
+  lightningNetworkSend: {
+    light: () => require('../svgJs/light/lightningNetworkSend-0').content,
+    dark: () => require('../svgJs/dark/lightningNetworkSend-0').content,
   },
   linkCoinbaseWallet: {
     light: () => require('../svgJs/light/linkCoinbaseWallet-3').content,
@@ -568,6 +568,10 @@ const svgJsMap = {
   referralsCoinbaseOne: {
     light: () => require('../svgJs/light/referralsCoinbaseOne-0').content,
     dark: () => require('../svgJs/dark/referralsCoinbaseOne-0').content,
+  },
+  referralsGenericCoin: {
+    light: () => require('../svgJs/light/referralsGenericCoin-0').content,
+    dark: () => require('../svgJs/dark/referralsGenericCoin-0').content,
   },
   retailUSDCRewards: {
     light: () => require('../svgJs/light/retailUSDCRewards-3').content,